+++ bonsai/src/bt.rs
@@ -1,5 +1,3 @@
-use std::fmt::Debug;
-
 use crate::{state::State, ActionArgs, Behavior, Float, Status, UpdateEvent};
 
 #[cfg(feature = "serde")]
@@ -11,40 +9,25 @@ use serde::{Deserialize, Serialize};
 #[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
 pub struct BT<A, B> {
     /// constructed behavior tree
-    state: State<A>,
+    pub(crate) state: State<A>,
     /// keep the initial state
-    initial_behavior: Behavior<A>,
+    pub(crate) initial_behavior: Behavior<A>,
     /// The data storage shared by all nodes in the tree. This is generally
     /// referred to as a "blackboard". State is written to and read from a
     /// blackboard, allowing nodes to share state and communicate each other.
-    bb: B,
+    pub(crate) bb: B,
     /// Whether the tree has been finished before.
-    finished: bool,
+    pub(crate) finished: bool,
     /// Monotonically increasing per-tick counter. Starts at 0; first completed
     /// `tick`/`tick_recording` call increments to 1. Survives `reset_bt`
     /// (the counter is global to the BT instance, not the current run).
-    tick_count: u64,
-    /// Preorder node metadata, computed once at `BT::new`.
-    /// Used by `RecordingTracer` to advance past unvisited subtrees in O(1).
+    pub(crate) tick_count: u64,
+    /// Bundle of visualize-only state: preorder node metadata, telemetry
+    /// channel sender, dropped-trace counter, and the per-tick recording
+    /// buffer. See [`crate::telemetry_state::TelemetryState`].
     #[cfg(feature = "visualize")]
     #[cfg_attr(feature = "serde", serde(skip))]
-    pub(crate) node_metas: Vec<crate::tracer::NodeMeta>,
-    /// Channel sender for shipping `TickTrace`s to the broadcaster thread.
-    /// `None` when telemetry is not active or after the broadcaster has exited.
-    #[cfg(feature = "visualize")]
-    #[cfg_attr(feature = "serde", serde(skip))]
-    pub(crate) telemetry_sender: Option<std::sync::mpsc::SyncSender<crate::telemetry::TickTrace>>,
-    /// Number of `TickTrace`s dropped because the channel was full.
-    /// Reset on `reset_bt`. Useful for diagnosing slow clients.
-    #[cfg(feature = "visualize")]
-    #[cfg_attr(feature = "serde", serde(skip))]
-    pub(crate) dropped_traces: u64,
-    /// Reusable buffer for the recording trace. Held for the BT's lifetime;
-    /// `tick_recording` clears it on entry, preserving capacity. Avoids one
-    /// `HashMap` allocation per tick on the hot path
-    #[cfg(feature = "visualize")]
-    #[cfg_attr(feature = "serde", serde(skip))]
-    pub(crate) trace_buffer: crate::telemetry::TickTrace,
+    pub(crate) telemetry: crate::telemetry_state::TelemetryState,
 }
 
 impl<A: Clone, B> BT<A, B> {
@@ -53,7 +36,8 @@ impl<A: Clone, B> BT<A, B> {
         let bt = State::new(behavior);
 
         #[cfg(feature = "visualize")]
-        let node_metas = crate::tracer::build_node_metas(&backup_behavior);
+        let telemetry =
+            crate::telemetry_state::TelemetryState::new(crate::telemetry::build_node_metas(&backup_behavior));
 
         Self {
             state: bt,
@@ -62,16 +46,7 @@ impl<A: Clone, B> BT<A, B> {
             finished: false,
             tick_count: 0,
             #[cfg(feature = "visualize")]
-            node_metas,
-            #[cfg(feature = "visualize")]
-            telemetry_sender: None,
-            #[cfg(feature = "visualize")]
-            dropped_traces: 0,
-            #[cfg(feature = "visualize")]
-            trace_buffer: crate::telemetry::TickTrace {
-                tick_id: 0,
-                states: std::collections::HashMap::new(),
-            },
+            telemetry,
         }
     }
 
@@ -98,26 +73,69 @@ impl<A: Clone, B> BT<A, B> {
         if self.finished {
             return None;
         }
-        // When telemetry is attached, delegate to the recording path so the
-        // broadcaster receives a TickTrace. One runtime branch; both sides
-        // keep their Tracer monomorphizations.
-        #[cfg(feature = "visualize")]
-        if self.telemetry_sender.is_some() {
-            return self.tick_recording(e, f).map(|(result, _)| result);
+        if let Some(out) = self.try_route_recording(e, f) {
+            return out;
         }
         self.tick_count += 1;
+        let result = self.dispatch_noop_tick(e, f);
+        if matches!(result, (Status::Success | Status::Failure, _)) {
+            self.finished = true;
+        }
+        Some(result)
+    }
+
+    /// Run `State::tick` with a [`NoopTracer`](crate::tracer::NoopTracer) (the
+    /// non-recording path). The cfg-gated `metas` binding lives inside this
+    /// helper so `tick`'s body can stay free of `#[cfg]` directives.
+    ///
+    /// Disjoint-field borrows: `&self.telemetry.node_metas` (immutable) and
+    /// `&mut self.state` / `&mut self.bb` (mutable) target distinct fields,
+    /// so the borrow checker accepts the simultaneous borrows.
+    ///
+    /// `#[inline(always)]` ensures the cfg branches constant-fold at the
+    /// monomorphization site, leaving identical generated code to the prior
+    /// inlined-in-`tick` version.
+    #[inline(always)]
+    fn dispatch_noop_tick<E, F>(&mut self, e: &E, f: &mut F) -> (Status, Float)
+    where
+        E: UpdateEvent,
+        F: FnMut(ActionArgs<E, A>, &mut B) -> (Status, Float),
+    {
         let mut tracer = crate::tracer::NoopTracer;
         #[cfg(feature = "visualize")]
-        let metas: &[crate::tracer::NodeMeta] = &self.node_metas;
+        let metas: &[crate::tracer::NodeMeta] = &self.telemetry.node_metas;
         #[cfg(not(feature = "visualize"))]
         let metas: &[crate::tracer::NodeMeta] = &[];
-        match self.state.tick(0, metas, e, &mut self.bb, f, &mut tracer) {
-            result @ (Status::Success | Status::Failure, _) => {
-                self.finished = true;
-                Some(result)
-            }
-            result => Some(result),
+        self.state.tick(0, metas, e, &mut self.bb, f, &mut tracer)
+    }
+
+    /// If telemetry is attached, dispatch to `tick_recording` and return its
+    /// result wrapped in `Some`. Returns `None` otherwise — the caller
+    /// (`tick`) should proceed with the no-op path.
+    ///
+    /// Returns `Option<Option<(Status, Float)>>`:
+    /// - `None` — helper did not handle the tick; caller continues.
+    /// - `Some(None)` — helper handled it but the BT was already finished.
+    /// - `Some(Some(result))` — helper handled it; this is the tick's return.
+    ///
+    /// `#[inline(always)]` lets the optimizer constant-fold the no-op path:
+    /// under `not(feature = "visualize")` the body is unconditionally `None`,
+    /// so the `if let Some(_) = self.try_route_recording(...)` branch in
+    /// `tick` becomes unreachable and disappears.
+    #[inline(always)]
+    fn try_route_recording<E, F>(&mut self, e: &E, f: &mut F) -> Option<Option<(Status, Float)>>
+    where
+        E: UpdateEvent,
+        F: FnMut(ActionArgs<E, A>, &mut B) -> (Status, Float),
+    {
+        #[cfg(feature = "visualize")]
+        if self.telemetry.sender.is_some() {
+            return Some(self.tick_recording(e, f).map(|(result, _)| result));
         }
+        // Suppress unused-variable warnings on the no-op path (visualize off,
+        // or visualize on but no sender attached).
+        let _ = (e, f);
+        None
     }
 
     /// Retrieve an immutable reference to the blackboard for
@@ -152,7 +170,7 @@ impl<A: Clone, B> BT<A, B> {
         // dropped_traces resets per-run: it's a diagnostic for the current session.
         #[cfg(feature = "visualize")]
         {
-            self.dropped_traces = 0;
+            self.telemetry.dropped_traces = 0;
         }
     }
 
@@ -165,244 +183,5 @@ impl<A: Clone, B> BT<A, B> {
     /// [`Status::Success`] or [`Status::Failure`]).
     pub fn is_finished(&self) -> bool {
         self.finished
-    }
-}
-
-
-#[cfg(feature = "visualize")]
-impl<A: Clone, B> BT<A, B> {
-    /// Number of `TickTrace`s dropped because the broadcaster channel was full.
-    /// Reset on `reset_bt`. Useful for diagnosing slow visualizer clients.
-    /// Always returns 0 if the visualizer was never attached via `BT::with_telemetry`.
-    pub fn dropped_traces(&self) -> u64 {
-        self.dropped_traces
-    }
-
-    /// Tick the tree once and return both the standard tick result and a
-    /// [`TickTrace`](crate::telemetry::TickTrace) recording every node visited
-    /// this frame. Used by the in-process telemetry channel and by integration
-    /// tests.
-    ///
-    /// Returns `None` if the tree has already finished (mirroring [`tick`](Self::tick)).
-    pub fn tick_recording<E, F>(
-        &mut self,
-        e: &E,
-        f: &mut F,
-    ) -> Option<((Status, Float), crate::telemetry::TickTrace)>
-    where
-        E: UpdateEvent,
-        F: FnMut(ActionArgs<E, A>, &mut B) -> (Status, Float),
-    {
-        if self.finished {
-            return None;
-        }
-        self.tick_count += 1;
-        // Reuse the long-lived buffer instead of fresh-allocating per tick.
-        // `clear()` preserves the HashMap's capacity, so once warmed up the
-        // fill phase doesn't reallocate.
-        self.trace_buffer.tick_id = self.tick_count;
-        self.trace_buffer.states.clear();
-        let result = {
-            let mut tracer = crate::telemetry::RecordingTracer {
-                trace: &mut self.trace_buffer,
-                metas: &self.node_metas,
-            };
-            self.state
-                .tick(0, &self.node_metas, e, &mut self.bb, f, &mut tracer)
-        };
-        if matches!(result, (Status::Success | Status::Failure, _)) {
-            self.finished = true;
-        }
-        // Try to ship the trace to the broadcaster thread. Uses as_ref().map() to
-        // release the immutable borrow before the match arms take mutable borrows.
-        if let Some(outcome) = self
-            .telemetry_sender
-            .as_ref()
-            .map(|tx| tx.try_send(self.trace_buffer.clone()))
-        {
-            use std::sync::mpsc::TrySendError;
-            match outcome {
-                Ok(()) => {}
-                Err(TrySendError::Full(_)) => self.dropped_traces += 1,
-                Err(TrySendError::Disconnected(_)) => self.telemetry_sender = None,
-            }
-        }
-        Some((result, self.trace_buffer.clone()))
-    }
-
-    /// Attach a live visualizer at `http://127.0.0.1:{port}/`.
-    ///
-    /// Convenience for [`with_telemetry_at`](Self::with_telemetry_at) with the
-    /// loopback address. Use that method instead if you need to bind a
-    /// different host (e.g. `"0.0.0.0"` to expose the visualizer on the LAN).
-    ///
-    /// Builder method: consumes and returns `Self` for chaining off [`BT::new`].
-    /// Spawns a listener thread and a broadcaster thread. Telemetry is
-    /// **best-effort** — [`TickTrace`](crate::telemetry::TickTrace)s are dropped
-    /// silently when the channel is full; the count is surfaced via
-    /// [`dropped_traces`](Self::dropped_traces).
-    ///
-    /// After calling this method, [`tick`](Self::tick) automatically records and
-    /// ships a trace on every call — no API change required.
-    ///
-    /// # Errors
-    /// Returns `io::Error` if `127.0.0.1:{port}` cannot be bound.
-    ///
-    /// # Example
-    /// ```no_run
-    /// # use bonsai_bt::{Action, BT, Behavior::Sequence};
-    /// # use std::collections::HashMap;
-    /// let behavior = Sequence(vec![Action("step")]);
-    /// let bb: HashMap<String, i32> = HashMap::new();
-    /// let mut bt = BT::new(behavior, bb).with_telemetry(8910)?;
-    /// # Ok::<(), std::io::Error>(())
-    /// ```
-    pub fn with_telemetry(self, port: u16) -> std::io::Result<Self>
-    where
-        A: std::fmt::Debug,
-    {
-        self.with_telemetry_at("127.0.0.1", port)
-    }
-
-    /// Attach a live visualizer at `http://{addr}:{port}/`.
-    ///
-    /// Like [`with_telemetry`](Self::with_telemetry), but lets you pick the
-    /// bind address. Pass `"0.0.0.0"` to listen on every interface (so peers
-    /// on the LAN can connect), `"127.0.0.1"` for loopback only, or any
-    /// specific interface IP. The address is parsed by `TcpListener::bind`,
-    /// so hostnames and IPv6 literals (e.g. `"::1"`) also work.
-    ///
-    /// Spawns a listener thread and a broadcaster thread. Telemetry is
-    /// **best-effort** — [`TickTrace`](crate::telemetry::TickTrace)s are dropped
-    /// silently when the channel is full; the count is surfaced via
-    /// [`dropped_traces`](Self::dropped_traces).
-    ///
-    /// After calling this method, [`tick`](Self::tick) automatically records and
-    /// ships a trace on every call — no API change required.
-    ///
-    /// Calling either telemetry method a second time replaces the existing
-    /// sender; the previous broadcaster exits on its next `recv` (sees
-    /// `Disconnected`). If the same `(addr, port)` is still bound, the second
-    /// bind returns `AddrInUse`.
-    ///
-    /// # Errors
-    /// Returns `io::Error` if `{addr}:{port}` cannot be bound (invalid address,
-    /// `AddrInUse`, permission denied for low ports, etc.).
-    ///
-    /// # Edge cases
-    /// - **Cloning a telemetry-attached `BT` is not supported.** `BT` derives
-    ///   `Clone`, and `Option<SyncSender<_>>` clones share the underlying
-    ///   channel — both BTs would interleave their `TickTrace`s into the same
-    ///   broadcaster, producing nonsense in the visualizer. Either clone
-    ///   *before* attaching telemetry, or call `with_telemetry_at` again on
-    ///   the clone (with a different port).
-    /// - **`reset_bt` preserves telemetry.** The sender survives, `tick_count`
-    ///   keeps climbing, `dropped_traces` resets to 0. Connected browsers see
-    ///   no disruption.
-    /// - **Already-finished BT.** `with_telemetry_at` still works — the tree
-    ///   definition is sent to clients on connect — but no `TickTrace`s flow
-    ///   until [`reset_bt`](Self::reset_bt) is called (`tick` returns `None`).
-    /// - **Deserializing a BT.** `telemetry_sender` is `#[serde(skip)]`, so the
-    ///   deserialized BT runs without telemetry until a telemetry method is
-    ///   called again — the typical "fresh run" workflow.
-    /// - **Pre-connect tick backlog.** Traces queue in the channel (capacity
-    ///   1024) until the first WS client connects. If no one connects within
-    ///   1024 ticks, older traces are dropped (`dropped_traces` increments);
-    ///   the first connecting client sees the tree definition plus the *next*
-    ///   tick onward, not the backlog.
-    /// - **Binding `0.0.0.0`.** The URL printed to stdout is the address you
-    ///   passed; modern browsers map `http://0.0.0.0` to localhost, but for a
-    ///   peer to connect they need this machine's actual LAN IP.
-    ///
-    /// # Example
-    /// ```no_run
-    /// # use bonsai_bt::{Action, BT, Behavior::Sequence};
-    /// # use std::collections::HashMap;
-    /// let behavior = Sequence(vec![Action("step")]);
-    /// let bb: HashMap<String, i32> = HashMap::new();
-    /// // Expose to the LAN — anyone who can reach this host can view the tree.
-    /// let mut bt = BT::new(behavior, bb).with_telemetry_at("0.0.0.0", 8910)?;
-    /// # Ok::<(), std::io::Error>(())
-    /// ```
-    pub fn with_telemetry_at(mut self, addr: &str, port: u16) -> std::io::Result<Self>
-    where
-        A: std::fmt::Debug,
-    {
-        use crate::telemetry::{TickTrace, TreeDefinition};
-        use std::net::TcpListener;
-        use std::sync::mpsc::sync_channel;
-
-        let listener = TcpListener::bind((addr, port))?;
-        // local_addr reflects the resolved socket — useful when port=0 (kernel-
-        // assigned) or when the user passes a hostname rather than a literal IP.
-        let bound = listener.local_addr()?;
-        let definition = serde_json::to_string(&TreeDefinition::build(&self.initial_behavior))
-            .expect("TreeDefinition is always serializable");
-        let (tx, rx) = sync_channel::<TickTrace>(1024);
-        crate::visualizer_server::spawn_server(listener, definition, rx)?;
-        self.telemetry_sender = Some(tx);
-        println!("bonsai-bt visualizer: open http://{bound}/");
-        Ok(self)
-    }
-}
-
-#[cfg(feature = "visualize")]
-impl<A: Clone + Debug, B: Debug> BT<A, B> {
-    /// Compile the behavior tree into a [graphviz](https://graphviz.org/) compatible [DiGraph](https://docs.rs/petgraph/latest/petgraph/graph/type.DiGraph.html).
-    ///
-    /// ```rust
-    /// use std::collections::HashMap;
-    /// use bonsai_bt::{
-    ///     Behavior::{Action, Sequence, Wait, WaitForever, While},
-    ///     BT
-    /// };
-    ///
-    /// #[derive(Clone, Debug, Copy)]
-    /// pub enum Counter {
-    ///     // Increment accumulator.
-    ///     Inc,
-    ///     // Decrement accumulator.
-    ///     Dec,
-    /// }
-    ///
-    ///
-    /// // create the behavior
-    /// let behavior = While(Box::new(WaitForever), vec![Wait(0.5), Action(Counter::Inc), WaitForever]);
-    ///
-    /// let h: HashMap<String, i32> = HashMap::new();
-    /// let mut bt = BT::new(behavior, h);
-    ///
-    /// // produce a string DiGraph compatible with graphviz
-    /// // paste the contents in graphviz, e.g: https://dreampuf.github.io/GraphvizOnline/#
-    /// let g = bt.get_graphviz();
-    /// println!("{}", g);
-    /// ```
-    pub fn get_graphviz(&mut self) -> String {
-        self.get_graphviz_with_graph_instance().0
-    }
-
-    pub(crate) fn get_graphviz_with_graph_instance(
-        &mut self,
-    ) -> (String, petgraph::Graph<crate::visualizer::NodeType<A>, u32>) {
-        use crate::visualizer::NodeType;
-        use petgraph::dot::{Config, Dot};
-        use petgraph::Graph;
-
-        let behavior = self.initial_behavior.to_owned();
-
-        let mut graph = Graph::<NodeType<A>, u32, petgraph::Directed>::new();
-        let root_id = graph.add_node(NodeType::Root);
-
-        Self::dfs_recursive(&mut graph, behavior, root_id);
-
-        let digraph = Dot::with_config(&graph, &[Config::EdgeNoLabel]);
-        (format!("{:?}", digraph), graph)
-    }
-
-    /// Compiles the behavior tree into a JSON string representing the static hierarchy.
-    pub fn get_telemetry_definition(&self) -> String {
-        let definition = crate::telemetry::TreeDefinition::build(&self.initial_behavior);
-        serde_json::to_string_pretty(&definition)
-            .expect("TreeDefinition is always serializable")
     }
 }
