+++ bonsai/src/telemetry.rs
@@ -6,12 +6,13 @@ use std::collections::HashMap;
 
 use serde::{Deserialize, Serialize};
 
-/// Re-exports preserving the `bonsai_bt::telemetry::*` paths used by tests
-/// and downstream code. The implementations live in the always-on
-/// [`crate::tracer`] module.
-pub use crate::tracer::{build_node_metas, NodeMeta};
+/// `NodeMeta` lives in the always-on [`crate::tracer`] module because
+/// `State::tick`'s signature depends on it regardless of the `visualize`
+/// feature. Re-exported here so the public path `bonsai_bt::telemetry::NodeMeta`
+/// stays valid for downstream code.
+pub use crate::tracer::NodeMeta;
 
-use crate::tracer::{children_of, Tracer};
+use crate::tracer::Tracer;
 use crate::{Behavior, Status};
 
 pub struct RecordingTracer<'a> {
@@ -49,6 +50,48 @@ pub struct TreeNode {
     pub node_type: &'static str,
     pub label: String,
     pub children: Vec<TreeNode>,
+}
+
+/// Walk `behavior` in DFS preorder and build a flat `Vec<NodeMeta>` indexed by
+/// preorder ID.  The ordering matches `TreeDefinition::traverse` exactly because
+/// both call `children_of`.
+pub fn build_node_metas<A>(behavior: &Behavior<A>) -> Vec<NodeMeta> {
+    let mut metas = Vec::new();
+    fill(behavior, &mut metas);
+    metas
+}
+
+fn fill<A>(b: &Behavior<A>, out: &mut Vec<NodeMeta>) -> usize {
+    let my_idx = out.len();
+    out.push(NodeMeta { subtree_size: 0 }); // placeholder, updated below
+    let mut size = 1;
+    for c in children_of(b) {
+        size += fill(c, out);
+    }
+    out[my_idx].subtree_size = size;
+    size
+}
+
+/// Returns the ordered children of a behavior node.
+///
+/// This is the **single source of truth** for preorder ID assignment order.
+/// `build_node_metas` and `TreeDefinition::traverse` must call this rather
+/// than re-implementing the ordering independently.
+pub(crate) fn children_of<A>(b: &Behavior<A>) -> Vec<&Behavior<A>> {
+    use Behavior::*;
+    match b {
+        Action(_) | Wait(_) | WaitForever => vec![],
+        Invert(c) | AlwaysSucceed(c) => vec![c.as_ref()],
+        // [condition, on_success, on_failure] — must match skip_subtree logic.
+        If(cond, ok, ko) => vec![cond.as_ref(), ok.as_ref(), ko.as_ref()],
+        While(cond, body) | WhileAll(cond, body) => {
+            let mut v = Vec::with_capacity(1 + body.len());
+            v.push(cond.as_ref());
+            v.extend(body.iter());
+            v
+        }
+        Select(xs) | Sequence(xs) | WhenAll(xs) | WhenAny(xs) | After(xs) | Race(xs) => xs.iter().collect(),
+    }
 }
 
 /// Returns the static node-type name and an optional dynamic label.
@@ -102,3 +145,77 @@ impl TreeDefinition {
 /// Embedded HTML payload served at `GET /` by the visualizer server.
 #[allow(dead_code)]
 pub const VISUALIZER_HTML: &str = include_str!("index.html");
+
+#[cfg(test)]
+mod tests {
+    use super::children_of;
+    use crate::Behavior::{self, Action, AlwaysSucceed, If, Invert, Select, Sequence, Wait, WaitForever, While};
+
+    #[derive(Clone, Debug)]
+    enum Act {
+        A,
+        B,
+        C,
+    }
+
+    fn ptrs<A>(xs: &[&Behavior<A>]) -> Vec<*const Behavior<A>> {
+        xs.iter().map(|b| *b as *const _).collect()
+    }
+
+    #[test]
+    fn children_of_leaves_yield_empty() {
+        assert!(children_of(&Action(Act::A)).is_empty());
+        assert!(children_of(&Wait::<Act>(1.0)).is_empty());
+        assert!(children_of(&WaitForever::<Act>).is_empty());
+    }
+
+    #[test]
+    fn children_of_if_is_cond_ok_ko() {
+        let cond: Box<Behavior<Act>> = Box::new(Action(Act::A));
+        let ok = Box::new(Action(Act::B));
+        let ko = Box::new(Action(Act::C));
+        let (cp, op, kp) = (&*cond as *const _, &*ok as *const _, &*ko as *const _);
+        let b = If(cond, ok, ko);
+        assert_eq!(ptrs(&children_of(&b)), vec![cp, op, kp], "If must be [cond, ok, ko]");
+    }
+
+    #[test]
+    fn children_of_while_is_cond_then_body() {
+        let cond: Box<Behavior<Act>> = Box::new(Action(Act::A));
+        let body0 = Action(Act::B);
+        let body1 = Action(Act::C);
+        let cp = &*cond as *const _;
+        let b = While(cond, vec![body0, body1]);
+        let kids = children_of(&b);
+        assert_eq!(kids.len(), 3, "While(cond, [B, C]) has 3 children");
+        assert_eq!(kids[0] as *const _, cp, "first child must be the condition");
+    }
+
+    #[test]
+    fn children_of_decorators_yield_single_child() {
+        let inner = Box::new(Action(Act::A));
+        let ip = &*inner as *const _;
+        let b = Invert(inner);
+        assert_eq!(ptrs(&children_of(&b)), vec![ip]);
+
+        let inner2 = Box::new(Action(Act::A));
+        let ip2 = &*inner2 as *const _;
+        let b2 = AlwaysSucceed(inner2);
+        assert_eq!(ptrs(&children_of(&b2)), vec![ip2]);
+    }
+
+    #[test]
+    fn children_of_composites_preserve_order() {
+        let items = vec![Action(Act::A), Action(Act::B), Action(Act::C)];
+        let seq = Sequence(items.clone());
+        let sel = Select(items.clone());
+        let seq_kids = children_of(&seq);
+        let sel_kids = children_of(&sel);
+        assert_eq!(seq_kids.len(), 3);
+        assert_eq!(sel_kids.len(), 3);
+        for i in 0..3 {
+            assert_eq!(format!("{:?}", seq_kids[i]), format!("{:?}", &items[i]), "Sequence child {i}");
+            assert_eq!(format!("{:?}", sel_kids[i]), format!("{:?}", &items[i]), "Select child {i}");
+        }
+    }
+}
